Updates on all pages
HTML - updated Class to be able to put the poem box as hidden, and then changing it via JS to be visible once the poem is generated.
JS - To give credits to the poem and change some structure to be able to have both, answer and credits. I had to create a new const
CSS - giving style for it to be more styled